Set permissions for /app folder only on first run and for some folders
The chown for the /app folder was very time expensive. It has no been changed
to only be executed if there is no LocalConfiguration.php which means (in most
cases) it is the first run of the container.
It now also only sets the permissions for the user-content folders which must
be writable by the web server.