[BUGFIX] Environmental variables not set in run.sh
The DB_HOST and DB_PORT shell variables will not get passed on to
the run.sh script chained with the bash "&&" operator. For this
reason the DB_HOST and DB_PORT environment variables were not
accessible to the "getenv()" in AdditionalConfiguration.php

The solution is to only execute "run-typo3.sh" from within the
Dockerfile and start "run.sh" from within "run-typo3.sh". By this
change the "run.sh" gets executed from within the environment
context of "run-typo3.sh" in which the mentioned variables are set.